Add files via upload
In this file there exists two types of codes for solving the question. The first one is inefficient and checks all penthagon numbers, the latter one is the most efficient and well-written code to check if each number is penthagonal or not. There is a lot to learn from the second code and several technics to tackle. In general, a nice elementary code.